Generate correct path for recent Torbrowser installs
TorBrowser is most commonly installed through the Torbrowser Launcher tool. 
This tool downloads a .tar.gz package and installs it in a directory inside the users $HOME. 

For quite some time, this directory is not ~/.torbrowser anymore,
but now the path varies depending on the system architecture and locale.

This change finds the actual folder to install the .json to.